requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117
Pillow==9.3.0
requests
tqdm==4.64.1
open_clip_torch==2.16.0
numpy
git+https://git@github.com/seatgeek/thefuzz.git@0.19.0#egg=thefuzz
python-Levenshtein==0.21.0